Bug fix/rocksdb truncate (#4328)
